System Overhaul 3: The domino effect was much larger than I anticipated, a lot of old code was coupled together had to be decoupled.
- Added new TextToSpeech to Warehouse NPCs.
- Food status icon now shows health regen amount per second.
- Gui highlighter for tutorials now shows text for improved communication on what to do.
- TheHarbor elevator reworked to new interactable system.
- New TaskHud interface for showing general tasks, for pinned mission and game mode tasks for Raids.
- Projectiles now use new projectilePackage format.
- New interactable system that is partially backwards compatible with old system, which allows NPC access to interact.
- New status system now works with NPCs with a status component, e.g. NPCs can be knocked down and revived by players.
- Added more world cores.
- Weapons interface now uses new UI engine.
- Fixed Melee ToolHandler and added OnPrimaryMelee modifier processor.
- Migrated all GameModeManager modules to RotdDataModel.
- Rewrote GameLobbyInterface as a `InterfaceWindow` for new UI engine.

I started an avalanche of overhauling old systems that were so coupled together because I wanted to rework some systems that really made it hard to add new mechanics and that spiraled to 7 months of development.
I wasn't initially expecting it to take that long and couple months in, I realize it's a sunk cost fallacy that I have to deal with. There won't likely be a devlog for Rotd for a while, so I'm posting my thoughts here for anyone interested.
The new systems so far are much more flexible for me to work with and I've been testing them on the Almdes project, I may post a dev log there if I have something interesting to talk about.
For the past few months, there was no end in sight to when this overhaul will be done, but I'm starting to see the way to the peak of this mountain, so I really want to thank everyone for your patience.